RDKB-9184:  User is able to surf web prior captive portal
Reason for change:
During activation scenarios, captive portal was presented to user before hitting activation portal.
With redesign, needed some extra check to prevent this.

Test Procedure:
Verified fix on the unit where issue was reported.
/var/dnsmasq.conf should not have captive portal configuration when unit is not activated.

Risks: None

+++ arch/intel_usg/boards/arm_shared/scripts/network_response.sh
@@ -102,7 +102,7 @@ then
 	do
 
 		# Remove all response files if it already exists
-		if [ -e $RESPONSE ]
+		if [ -e $RESPONSE ] && [ -f "/tmp/.gotnetworkresponse" ]
 		then
 			responseCode=`cat $RESPONSE`
 			rm -rf $RESPONSE				
@@ -283,17 +283,19 @@ then
                               break
                            fi
                         fi
-                        echo_t "Network Response: Restarting v6Service"
-                        # Modify DNS server option in dibbler configuration
-                        if [ -e $SERVER6_CONF ]
-                        then
-                           sed -e '/dns-server/s/^/#/g' -i $SERVER6_CONF 
-                           cat $SERVER6_CONF
-                        else
-                           echo_t "No dibbler6 configuration available...."
-                        fi
-
                     done
+                    echo_t "Network Response: Restarting dibblerServer"
+                    # Modify DNS server option in dibbler configuration
+                    if [ -e $SERVER6_CONF ]
+                    then
+                        sed -e '/dns-server/s/^/#/g' -i $SERVER6_CONF 
+                        cat $SERVER6_CONF
+                    else
+                        echo_t "No dibbler6 configuration available...."
+                    fi
+                    
+                    dibbler-server stop
+                    dibbler-server start
                 else
                     echo_t "Network Response: WiFi is already personalized setting redirection_flag to false"
                     #We reached here as redirection_flag is "true". But WiFi is configured already as per notification status.
